Split out interpreter/eval.rs
